[regex] Moving keywords into a resource file
The main keywords regular expressions for macros, definitions,
clojure.core functions, etc. were starting to both (a) cause Vim to slow
down (wtf?) and (b) to make it impossible to actually read what was
in the buffer at any one time. This commit adds logic to quickly build a
regex for them based off the space-delimited keywords in the file and to
go from there.