[regex] First major working commit
This commit marks the first version of this library that seems to be
essentially fully functional. There's some stuff still to be done (e.g.
adding a README, CONTRIBUTING, CHANGELOG, various OSS stuff, etc.), but
otherwise it's more or less good to go.

Roadmap for 0.1.0 should include refactoring the core namespace to use a
macro rather than the ridiculously repetitive code structure currently
in use. Also, allowing for an options map that redefs the colorscheme.