[glow] Project cleanup
This commit marks the end of my long winter in the land of lexers and
grammars. I hope never to return here, or at least to only return here
when someone else is doing all of the heavy lifting.

Ultimately, ANTLR offers the key features that this project needs: it's
resilient, fast, accessible via Clojure (thanks @aphyr!), and outputs
a parse tree that is similar enough to what Instaparse uses that I can
leverage Instaparse's transforms for quick highlighting.

I've pulled out most of the old materials, including the original
regular expressions (prone to StackOverflowErrors due to Java 6's regex
implementation), and the Clojure EBNF I was working on with Instaparse
(too slow, still hits the same StackOverflowErrors).