[bugfix] Fixes parsing of single-line comment strings
The grammar currently only parses comments that end in a newline. If you
send a string of code that includes a comment but that doesn't include a
newline currently the parser barfs. This commit makes it so that an EOF
token is also acceptable and makes parsing single-line comment strings
possible.